Core OFRAK speed improvements (#61)
This should, in particular, speed up `resource.unpack_recursively` calls where `resource` is an ELF or other executable binary with code regions. Since there are few calls like that in the tests, it's likely the changeset will not significantly decrease the time required for CI pipeline tests. (Looks to be an improvement of roughly a minute for the main tests in my observations.)