GUI vaddr search (#70)
Adds a search functionality in the GUI, accessed from the toolbar. With a resource selected, one can search its descendants for resources with a VirtualAddress matching a specific value or within a range. These are displayed in the search view; each one can be selected as usual, and after going back to the main view, it remains selected.

Co-authored-by: edward <edward@redballoonsecurity.com>