Move away from "$OFRAK_DIR" argument to $PACKAGE_PATH for Docker builds
- $PACKAGE_PATH generalizes build_image.py such that it works with any
yaml file containing a valid list of Python packages

Before this change, it was possible to use build_image.py a directory above
the OFRAK repository by passing in an OFRAK_DIR arugment in the yml file.
This change generalizes the approach so that build_image.py can be run in any ancestor directory.