Make proper module for ofrak_gpu. test_entropy_component_gpu now works. no_gpu still needs work.
Set up ofrak_gpu to be a proper module. It now has a makefile, requerements, setup.py, etc. The dockerstub automatically installs the latest release of futhark and is used to create entropy.py from entropy.fut at build time. Unfortunately, this will probably get rejected, which sucks because making that Dockerstub was a pain.

Had to split up the entropy tests into two separate files, one that runs the tests using a fixture to always trick the system into thinking ofrak_gpu is installed, to test the DataSummaryAnalyzer's behavior when ofrak_gpu is present, and one that uses a fixture to always trick the system into thinking ofrak_gpu has not been installed, to ensure DataSummaryAnalyzer's current behavior remains unaffected. These fixtures were incredibly annoying and confusing to get working consistently. As of this commit, no_gpu is still in need of more research. However, these efforts will allow us to fully test DataSummaryAnalyzer's behavior consistently, regardless of whether or not the ofrak_gpu module has been installed on the testing environment, which is very important for reliability (and unforuntately very difficult to do).